Split main file into more files
